Remove unnecessary import of librosa
