+++ dali/test/python/test_audio_decoder_utils.py
@@ -1,6 +1,5 @@
 import numpy as np
 import math
-import librosa
 
 
 def generate_waveforms(length, frequencies):
@@ -18,18 +17,3 @@ def generate_waveforms(length, frequencies):
 
     wave = np.sin(X[:, np.newaxis] * (np.array(frequencies) * (2 * math.pi)))
     return wave * window(X)[:, np.newaxis]
-
-
-def rosa_resample(input, in_rate, out_rate):
-    if input.shape[1] == 1:
-        return librosa.resample(input[:, 0], orig_sr=in_rate, target_sr=out_rate)[:, np.newaxis]
-
-    channels = [
-        librosa.resample(np.array(input[:, c]), orig_sr=in_rate, target_sr=out_rate)
-        for c in range(input.shape[1])
-    ]
-    ret = np.zeros(shape=[channels[0].shape[0], len(channels)], dtype=channels[0].dtype)
-    for c, a in enumerate(channels):
-        ret[:, c] = a
-
-    return ret
